Fixes double free when exiting with a signal.
When a signal caused the program to exit and call the sig_cleanup
function, the exit call triggered the cleanup hook again so the
cleanup function was called twice. Hence the PCAP file being freed
twice. This patch removes the first call and relies on the exit
hook to trigger the cleanup function.