Adds a warning message when the input buffer is full.
This adds a warning message when the UART input buffer is full. When
such a case arise it doesn't mean that we will drop frames but instead
that the kernel will buffer them for us. However it is unlikely that
the kernel will have to drop anything but it may vary with the nature
of the stdout file descriptor. It would rather point out an error
in the embedded firmware.