Fixes the information frames being always 127 bytes long.
The helper function used in the firmware to send an information string
miscalculated the maximal length of the frame chunk to send over to the
client. This resulted in informations being sent from beyond the
firmware buffer which in most case was filled with zero hence displaying
no visible bug.