Adds code to setup line control.
When a speed is specified at the command line we also want to setup the
serial line control options. In most cases we would like to set this to
8N1. That is to say 8 data bits, no parity and 1 stop bit. This patch
set these options by default when the speed is specified at the command
line. Otherwise the user has to setup the line control options manually
with stty.