Fixes line control setup.
When setting the line up we retrieved the original options on which we
updated necessary flags. It often led to forgotten flag and a line which
was then not set up properly. In this patch we do not base ourselves on
the original options but instead start from a zeroed structure that we
fill with the only bits to set.