Adds a generic atoi.
This patch adds a generic atoi function which understand number written
in different bases. This function is used to parse command line
arguments but it may be used for other purposes too. However this
function is optimistic and does not check for any error that may occur
(overflow, invalid symbols) when converting the number.