Flush UART on opening.
Some operating systems (eg Linux) bufferise the UART input even when the
file descriptor is not opened. This may be useful in a lot of cases.
However we may be in a case where incomplete messages still lay on the
UART buffer. Therefore we have to flush the buffer ourself.