Wait for the firmware to be ready.
This adds support to wait for the firmware to be ready before starting
transmissions and operations. We achieve this with a first ready byte
that is send by the firmware when the protocol is set up and ready to
receive data. The client associate a timeout to this ready byte and
wait for it. Only once the open procesure has read the ready byte he
will return control to the client.