Fixes the firmware side of the protocol.
The buffer used in the firmware side of the protocol input step was
allocated on the stack. This caused the firmware to decode packet only
by chance when the message was small enough and the stack didn't change
since the last call. However larger frames were not decoded correctly.
This behaviour was highlighted with large ping requests. This patch
fixes this by allocating the buffer on the heap instead.