Waits for the processing of the message in the injector command.
When an ACK was not requested, the injector command exited immediatly
after the frame has been written to UART. There were not confirmation
that the frame sent was processed correctly by the firmware. This patch
fixes this by modifying the message callback into a tiny state machine
that will receive the processing confirmation.