Fixes the decoding of frames without payload.
When a frame contained an empty payload, the mac decoding function would
parse an element of negative size. This case is normally captured and
the payload is set to NULL. However this negative size was converted to an
unsigned int and taken into account for the frame size check. This
resulted in an incorrect frame size and an error from the mac decode
function.