Fix memory corruption with xatou(32|64).
In the asm template, we used err incorrectly as an output operand.
But it really is an input operand as we only use the pointer value.

As a result the compiler sometime rewrote the value of err pointer in
the asm template as it expected us to change its value anyway, hence
resulting in a segfault at best or weird memory corruption in most cases.